# apps/core-api/docs/swagger.yaml
basePath: /
definitions:
  controller.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/repo.User'
    type: object
  controller.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  controller.ProductsListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/repo.Product'
        type: array
    type: object
  controller.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
    type: object
  repo.CreateOrderInput:
    properties:
      customer_notes:
        type: string
      items:
        items:
          $ref: '#/definitions/repo.CreateOrderItemInput'
        type: array
    type: object
  repo.CreateOrderItemInput:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  repo.Order:
    properties:
      created_at:
        type: string
      currency:
        type: string
      customer_notes:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/repo.OrderItem'
        type: array
      status:
        type: string
      subtotal:
        type: number
      tax:
        type: number
      total:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  repo.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      product_sku:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
      unit_price:
        type: number
    type: object
  repo.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      is_active:
        type: boolean
      is_digital:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      updated_at:
        type: string
    type: object
  repo.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Core API for the online-storage system.
  title: Core API
  version: 0.1.0
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - auth
  /api/orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/repo.CreateOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repo.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /api/orders/{id}:
    get:
      parameters:
      - description: Order ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Order'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /api/products:
    get:
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductsListResponse'
      summary: List products
      tags:
      - products
  /api/products/{id}:
    get:
      parameters:
      - description: Product ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repo.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - products
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"